requirements.txt
pytest
allure-pytest
pytest-html